chore: refactored getUnreachableAnswersAsNull
