+++ src/main/app/case/answers/possibleAnswers.ts
@@ -1,7 +1,9 @@
 import { pick } from 'lodash';
 
+import { stepsWithContent } from '../../../steps';
 import { Step } from '../../../steps/sequence';
 import { FormContent, FormOptions } from '../../form/Form';
+import { CaseApi } from '../CaseApi';
 import { Case } from '../case';
 
 type StepWithForm = { form?: FormContent } & Step;
@@ -42,3 +44,12 @@ export const getAllPossibleAnswers = (caseState: Partial<Case>, steps: Step[]): 
 
 export const omitUnreachableAnswers = (caseState: Partial<Case>, steps: Step[]): Partial<Case> =>
   pick(caseState, getAllPossibleAnswers(caseState, steps));
+
+export const getUnreachableAnswersAsNull = (userCase: Partial<Case>): Partial<Case> => {
+  const possibleAnswers = getAllPossibleAnswers(userCase, stepsWithContent);
+  return Object.fromEntries(
+    Object.keys(userCase)
+      .filter(key => !CaseApi.READONLY_FIELDS.includes(key) && !possibleAnswers.includes(key) && userCase[key] !== null)
+      .map(key => [key, null])
+  );
+};
